requirements.txt
requests
pipedrive-python-lib
PyBambooHR
python-dateutil
sqlalchemy
dictalchemy
pymysql
